requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pynput>=1.7.6
keyboard>=0.13.5
pygetwindow>=0.0.9

